[REFACTOR] switch app.js dir location
