+++ uni_frontend/src/App.js
@@ -1,19 +1,19 @@
 import React, { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
-import Layout from '../../layout';
-import Mainpage from '../main-page/mainpage';
-import Register from '../register-page/Register';
-import Login from '../login-page/login';
-import UserPage from '../user-page/UserPage';
-import EditPage from "../edit-page/EditPage";
-import ChatPage from "../chat-page/ChatPage";
-import Admin from "../admin-page/admin";
+import Layout from './layout';
+import Mainpage from './pages/main-page/mainpage';
+import Register from './pages/register-page/Register';
+import Login from './pages/login-page/login';
+import UserPage from './pages/user-page/UserPage';
+import EditPage from "./pages/edit-page/EditPage";
+import ChatPage from "./pages/chat-page/ChatPage";
+import Admin from "./pages/admin-page/admin";
 import Cookies from "js-cookie";
-import MatchingStatus from '../navbar/matchingList';
-import ChatList from '../navbar/chatList';
-import Review from '../review-page/review';
-import Forget from "../../Forgat";
-import usePushNotification from "../../Alarm";
+import MatchingStatus from './pages/navbar/matchingList';
+import ChatList from './pages/navbar/chatList';
+import Review from './pages/review-page/review';
+import Forget from "./pages/login-page/Forgat";
+import usePushNotification from "./Alarm";
 import SockJS from "sockjs-client";
 import { Stomp } from "@stomp/stompjs";
