+++ uni_frontend/src/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './pages/style/App';
+import App from './App';
 
 test('renders learn react link', () => {
   render(<App />);
