+++ uni_frontend/src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from '../../reportWebVitals';
-import '../../locales/i18n';
+import reportWebVitals from './reportWebVitals';
+import './locales/i18n';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
